server/pyproject.toml
[tool.poetry]
name = "krakenci-server"
version = "0.0"
description = "Kraken CI server."
authors = ["Michal Nowikowski <godfryd@gmail.com>"]
readme = "README.md"
homepage = "https://kraken.ci/"
repository = "https://github.com/kraken-ci/kraken"
documentation = "https://kraken.ci/docs"
keywords = ["building", "testing", "continuous-integration", "ci", "cd", "cicd"]
license = "Apache-2.0"

classifiers = [
    "Topic :: Software Development",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "License :: OSI Approved :: Apache Software License"
]

packages = [
  { include = "kraken" },
]

include = ["kraken/version.py"]

[tool.poetry.dependencies]
python = "^3.7"
connexion = "^2.14.2"
MarkupSafe = "2.1.2"
Flask = "^2.2.3"
Flask-SQLAlchemy = "^3.0.3"
psycopg2-binary = "^2.9.1"
redis = "^3.5.3"
requests = "^2.26.0"
SQLAlchemy = "^1.4.25"
swagger-ui-bundle = "^0.0.9"
apscheduler = "^3.8.0"
tzlocal = "2.1"  # this is required by apscheduler that still relies on pytz; aps 4.x will have this fixed and then this line can be removed
pytimeparse = "^1.1.8"
python-dateutil = "^2.8.2"
gunicorn = "^20.1.0"
RestrictedPython = "5.0"
alembic = "^1.7.3"
giturlparse = "^0.10.0"
clickhouse-driver = "^0.2.5"
minio = "^7.1.0"
passlib = "^1.7.4"
sentry-sdk = {extras = ["flask"], version = "^1.5.0"}
boto3 = "^1.18.52"
jsonschema = "^4.5.0"
rq = "^1.10.0"
azure-identity = "^1.6.1"
azure-mgmt-compute = "^23.0.0"
azure-mgmt-monitor = "^2.0.0"
azure-mgmt-network = "^19.0.0"
azure-mgmt-resource = "^19.0.0"
azure-mgmt-storage = "^18.0.0"
azure-mgmt-subscription = "^1.0.0"
kubernetes = "^20.13.0"
casbin = "^1.17.1"
python-ldap = "^3.4.3"
Authlib = "^1.1.0"
setuptools = "^66.0.0"  # required by apscheduler
furl = "^2.1.3"
jq = "^1.4.0"
jsonpatch = "^1.32"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pylint = "2.13.9"
pudb = "^2021.1"
PyHamcrest = "^2.0.3"

[tool.poetry.scripts]
# kkserver = 'kraken.server.server:main' # this does not work due to flask problems
kkscheduler = 'kraken.server.scheduler:main'
kkplanner = 'kraken.server.planner:main'
kkwatchdog = 'kraken.server.watchdog:main'
kkrq = 'kraken.server.kkrq:main'
kkqneck = 'kraken.server.qneck:main'
kkdbmigrate = 'kraken.migrations.apply:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
